requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
bard
stripe==10.7.0
mysql-connector-python==9.0.0
telebot==0.0.5
youtube_dl==2021.12.17
bing_image_downloader 
beautifulsoup4
bs4
cloudscraper
deepai
daxxhub
dnspython
emojis
ffmpeg-python
Faker
future==1.0.0
git+https://github.com/AsmSafone/SafoneAPI
gitpython
gpytranslate
gtts
hachoir
heroku3
httpx==0.27.0
img2pdf
instaloader
lexica-api
motor 
#MukeshAPI
numpy
openai==0.28.0
opencv-python
pillow==9.5.0
psutil
pickledb
pornhub-api==0.3.0
pycountry
pycryptodome
pydub
pyfiglet
pykeyboard
python-dotenv==1.0.1
python-telegram-bot==21.1.1
python-whois
pyshorteners
pyttsx3
pytz
pytube
py-tgcalls==0.9.7
pyyaml==6.0.1
qrcode
requests
SafoneAPI
search_engine_parser
speedtest-cli
SpeechRecognition
spotipy
telegraph
tgcrypto==1.2.5
unidecode
wget
youtube-search
youtube-search-python
git+https://github.com/KurimuzonAkuma/pyrogram
git+https://github.com/yt-dlp/yt-dlp
